moved demo to home page
