+++ app/page.tsx
@@ -1,60 +1,39 @@
-import Image from 'next/image'
+'use client'
 
-export default function Home() {
-  return (
-    <main className="min-h-screen p-24">
-      <h1 className="text-2xl text-center mb-6">Demos</h1>
-      <p className="text-center mb-6">AKA some of the skill tests I have been asked to create.</p>
+import DropdownSelect from '../components/DropdownSelect';
 
-      <div className="mb-32 grid text-center lg:max-w-5xl lg:w-full lg:mb-0 lg:grid-cols-3 lg:text-left">
-        <a
-          href="demo/dropdownselect"
-          className="group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-gray-300 hover:bg-gray-100 hover:dark:border-neutral-700 hover:dark:bg-neutral-800/30"
+const DropdownDemo = () => {
+    const options = [
+        { value: 'red', label: 'Red' },
+        { value: 'green', label: 'Green' },
+        { value: 'blue', label: 'Blue' },
+        { value: 'purple', label: 'Purple' },
+        { value: 'yellow', label: 'Yellow' },
+        { value: 'white', label: 'White'},
+    ];
 
-        >
-          <h2 className={`mb-3 text-2xl font-semibold`}>
-            DropdownSelect{' '}
-            <span className="inline-block transition-transform group-hover:translate-x-1 motion-reduce:transform-none">
-              -&gt;
-            </span>
-          </h2>
-          <p className={`m-0 max-w-[30ch] text-sm opacity-50`}>
-            Reusable React dropdown component
-          </p>
-        </a>
+    return (
+        <main className="min-h-screen p-4">
+            <h1 className="mb-4 text-lg">DropdownSelect Reusable React Component Demo</h1>
+            <p className="mb-4"><a href="https://github.com/katymdc/portfolio-website" className="underline text-blue-600">Git Source Code</a></p>
+            <p className="mb-1">Technologies Used</p>
+            <ul className="mb-8 ml-8 list-disc">
+                <li>ReactJS</li>
+                <li>AWS (S3, Route 53, CloudFront)</li>
+                <li>NextJS *NEW*</li>
+                <li>Tailwind CSS *NEW*</li>
+                <li>TypeScript *NEW*</li>
+                <li>Visual Studio Code *NEW*</li>
+            </ul>
 
-        <a
-          href="demo/pancakes"
-          className="group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-gray-300 hover:bg-gray-100 hover:dark:border-neutral-700 hover:dark:bg-neutral-800/30"
-        >
-          <h2 className={`mb-3 text-2xl font-semibold`}>
-            Pancakes{' '}
-            <span className="inline-block transition-transform group-hover:translate-x-1 motion-reduce:transform-none">
-              -&gt;
-            </span>
-          </h2>
-          <p className={`m-0 max-w-[30ch] text-sm opacity-50`}>
-            Coming soon
-          </p>
-        </a>
-
-        <a
-          href="demo/angularform"
-          className="group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-gray-300 hover:bg-gray-100 hover:dark:border-neutral-700 hover:dark:bg-neutral-800/30"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <h2 className={`mb-3 text-2xl font-semibold`}>
-            Angular Form{' '}
-            <span className="inline-block transition-transform group-hover:translate-x-1 motion-reduce:transform-none">
-              -&gt;
-            </span>
-          </h2>
-          <p className={`m-0 max-w-[30ch] text-sm opacity-50`}>
-            We will have to see about this one...
-          </p>
-        </a>
-      </div>
-    </main>
-  )
+            <DropdownSelect
+                label="Select a color"
+                options={options}
+                onChange={(value: string) => console.log(value)}
+                isSearchable
+            />
+        </main>
+    );
 }
+
+export default DropdownDemo;
